% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_autism_de_novos.R
\name{autism_de_novos}
\alias{autism_de_novos}
\title{get de novo data from autism exome studies}
\usage{
autism_de_novos()
}
\value{
data frame of de novos, including gene symbol, functional consequence
    (VEP format), chromosome, nucleotide position and SNV or INDEL type
}
\description{
I think this data was obtained from studies published using the Simon's
Simplex Collection data (http://sfari.org/resources/simons-simplex-collection)
}
\details{
Supplementary table 2 (where the excel sheets for the probands and
siblings have been combined) from:
Sanders et al. (2012) Nature 485:237-241
doi: 10.1038/nature10945

Supplementary table 3 from:
O'Roak et al. (2012) Nature 485:246-250
doi: 10.1038/nature10989

Supplementary table 1 (where the non-coding SNVs have been excluded) and
supplementary table 2 from:
Iossifov et al. (2012) Neuron 74:285-299
doi: 10.1016/j.neuron.2012.04.009
}

