% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_de_rubeis_de_novos.R
\name{de_rubeis_de_novos}
\alias{de_rubeis_de_novos}
\title{get de novo data from the 2014 De Rubeis et al. autism exome study in Nature}
\usage{
de_rubeis_de_novos()
}
\value{
data frame of de novos, including gene symbol, functional consequence
    (VEP format), chromosome, nucleotide position and SNV or INDEL type
}
\description{
De novo mutation data sourced from Supplementary table 3:
De Rubeis et al. (2013) Nature 515:209-215
doi: 10.1038/nature13772
}

