% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_de_ligt_de_novos.R
\name{deligt_de_novos}
\alias{deligt_de_novos}
\title{get de novo data for De Ligt et al. intellectual disability exome study}
\usage{
deligt_de_novos()
}
\value{
data frame of de novos, including gene symbol, functional consequence
    (VEP format), chromosome, nucleotide position and SNV or INDEL type
}
\description{
De novo mutation data sourced from supplementary table 3 from
De Ligt et al. (2012) N Engl J Med 367:1921-1929
doi: 10.1056/NEJMoa1206524
}

