% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_epi4k_de_novos.R
\name{epi4k_de_novos}
\alias{epi4k_de_novos}
\title{get de novo data for the Epi4K epilepsy exome study}
\usage{
epi4k_de_novos()
}
\value{
data frame of de novos, including gene symbol, functional consequence
    (VEP format), chromosome, nucleotide position and SNV or INDEL type
}
\description{
De novo mutation data from the most recent EPI4K publication:
Supplementary table 1:
American Journal of Human Genetics (2014) 95:360-370
doi: 10.1016/j.ajhg.2014.08.013
}
\details{
This incorporates the de novo mutation data from supplementary table 2 of:
Allen et al. (2013) Nature 501:217-221
doi: 10.1038/nature12439
}

