% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{find_matching_sites}
\alias{find_matching_sites}
\title{this identifies variants which match the current row}
\usage{
find_matching_sites(row, de_novos)
}
\arguments{
\item{row}{dataframe of a single row for a candidate, with columns for
person_id, chrom, start_pos, ref_allele}

\item{de_novos}{dataframe of all variants}
}
\value{
boolean vector for sites that match the current row (the current
        row should match at minimum).
}
\description{
Some variants in the same individual changed coordinates between studies.
These are all indels, so are due to realignment calling the variant at
slightly different locations. We can identify the initial position, by
finding the variant that is closest to the validation coordinates.
}
\examples{
sample = read.table(header=TRUE, text="
    person_id    chrom start_pos ref_allele alt_allele consequence study_phenotype
    sample_01    1     10000     AAA        A          frameshift_variant autism",
    stringsAsFactors=FALSE)

variants = read.table(header=TRUE, text="
    person_id    chrom start_pos ref_allele alt_allele consequence study_phenotype
    sample_01    1     10000     AAA        A          frameshift_variant autism
    sample_01.p1 1     10001     AAA        A          frameshift_variant normal_iq_autism
    sample_02    1     10001     AAA        A          frameshift_variant normal_iq_autism",
    stringsAsFactors=FALSE)
find_matching_sites(sample, variants)
}

