% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_fromer_de_novos.R
\name{fromer_de_novos}
\alias{fromer_de_novos}
\title{get de novo data from Fromer et al. schizophrenia exome study}
\usage{
fromer_de_novos()
}
\value{
data frame of de novos, including gene symbol, functional consequence
    (VEP format), chromosome, nucleotide position and SNV or INDEL type
}
\description{
De novo mutation data sourced from Supplementary table 1:
Fromer et al. (2014) Nature 506:179-184
doi: 10.1038/nature12929
}

