% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_de_ligt_de_novos.R
\name{get_deligt_denovo_table}
\alias{get_deligt_denovo_table}
\title{load the supplementary table for the de ligt de novo variants}
\usage{
get_deligt_denovo_table(pdf)
}
\arguments{
\item{pdf}{list of text lines in supplementary pdf.}
}
\value{
list of text in supplementary file
}
\description{
load the supplementary table for the de ligt de novo variants
}

