% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_de_ligt_de_novos.R
\name{get_deligt_sex}
\alias{get_deligt_sex}
\title{load the supplementary table for the de ligt sex information}
\usage{
get_deligt_sex(pdf)
}
\arguments{
\item{pdf}{list of text lines in supplementary pdf.}
}
\value{
dataframe of person_ids and sex codes
}
\description{
load the supplementary table for the de ligt sex information
}

