% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_de_ligt_de_novos.R
\name{get_deligt_text}
\alias{get_deligt_text}
\title{load the supplementary data for the de ligt de novos}
\usage{
get_deligt_text(url)
}
\arguments{
\item{url}{url of the de ligt supplementary pdf.}
}
\value{
list of text in supplementary file
}
\description{
load the supplementary data for the de ligt de novos
}

