% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gilissen_de_novos.R
\name{gilissen_de_novos}
\alias{gilissen_de_novos}
\title{get de novo data for Gilissen et al. intellectual disability exome study}
\usage{
gilissen_de_novos(deligt, deligt_sex)
}
\arguments{
\item{deligt}{dataframe of de novos variants from the De Ligt N Engl J Med
367:1921-1929 study, so we can get the sample IDs and sex for the
current dataset.}

\item{deligt_sex}{dataframe of sample IDs and sex codes for all of the
children in the de ligt study, not just those who had de novo variants.}
}
\value{
data frame of de novos, including gene symbol, functional consequence
    (VEP format), chromosome, nucleotide position and SNV or INDEL type
}
\description{
De novo mutation data sourced from supplementary table 8 from:
Gilissen et al. (2014) Nature 511:344-347
doi: 10.1038/nature13394
}
\details{
NOTE: the number of males and females will form part of the De Ligt counts,
NOTE: since the Gilissen samples were in the De Ligt study, they just didn't
receive a diagnosis in the original study
}

