% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_iossifov_nature_de_novos.R
\name{iossifov_nature_de_novos}
\alias{iossifov_nature_de_novos}
\title{get de novo data from the 2014 Iossifov et al. autism exome study in Nature}
\usage{
iossifov_nature_de_novos()
}
\value{
data frame of de novos, including gene symbol, functional consequence
    (VEP format), chromosome, nucleotide position and SNV or INDEL type
}
\description{
De novo mutation data sourced from Supplementary table 2:
Iossifov et al. (2014) Nature 498:216-221
doi: 10.1038/nature13908
}

