% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_iossifov_neuron_de_novos.R
\name{iossifov_neuron_de_novos}
\alias{iossifov_neuron_de_novos}
\title{get de novo data from the 2012 Iossifov et al autism exome study in Neuron}
\usage{
iossifov_neuron_de_novos()
}
\value{
data frame of de novos, with standardised genome coordinates and VEP
}
\description{
Supplementary table 1 (where the non-coding SNVs have been excluded) and
supplementary table 2 from:
Iossifov et al. (2012) Neuron 74:285-299
doi: 10.1016/j.neuron.2012.04.009
}

