% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lelieveld_de_novos.R
\name{lelieveld_de_novos}
\alias{lelieveld_de_novos}
\title{get de novo data for Lelieveld et al. intellectual disability exome study}
\usage{
lelieveld_de_novos()
}
\value{
data frame of de novos, including gene symbol, functional consequence
    (VEP format), chromosome, nucleotide position and SNV or INDEL type
}
\description{
De novo mutation data sourced from supplementary table 1 from:
Lelieveld et al. (2016) Nature Neuroscience 19:1194-1196
doi: 10.1038/nn.4352
}
\details{
Note that the paper says that the variants were aligned to hg19, but their
table of de novo variants is definitely for hg18 (GRCh37).
}

