% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_rest.R
\name{map_coordinates}
\alias{map_coordinates}
\title{map coordiantes from one assembly to another}
\usage{
map_coordinates(variant, inbuild = "grch38", outbuild = "grch37",
  build = "grch37", verbose = FALSE)
}
\arguments{
\item{variant}{data frame or list for a variant, containing columns named
"chrom", "start_pos", and "end_pos" for a single variant}

\item{inbuild}{genome build to find consequences on}

\item{outbuild}{genome build to find consequences on}

\item{build}{build of REST server to use}

\item{verbose}{flag indicating whether to print variants as they are checked}
}
\value{
a character string containing the HGNC symbol.
}
\description{
map coordiantes from one assembly to another
}
\examples{
map_coordinates(data.frame(chrom=c("1"), start_pos=c("1000000"),
    end_pos=c("1000000")))
map_coordinates(list(chrom="1", start_pos="1000000",
    end_pos="1000000"))
}

