% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_oroak_de_novos.R
\name{oroak_de_novos}
\alias{oroak_de_novos}
\title{get de novo data from the O'Roak et al autism exome study}
\usage{
oroak_de_novos()
}
\value{
data frame of de novos, with standardised genome coordinates and VEP
}
\description{
Supplementary table 3 from:
O'Roak et al. (2012) Nature 485:246-250
doi: 10.1038/nature10989
}

