% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rauch_de_novos.R
\name{rauch_de_novos}
\alias{rauch_de_novos}
\title{get de novo data for Rauch et al. intellectual disability exome study}
\usage{
rauch_de_novos()
}
\value{
data frame of de novos, including gene symbol, functional consequence
    (VEP format), chromosome, nucleotide position and SNV or INDEL type
}
\description{
De novo mutation data sourced from supplementary tables 2 and 3 from
Rauch et al. (2012) Lancet 380:1674-1682
doi: 10.1016/S0140-6736(12)61480-9
}

