% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{remove duplicate variants from the same individual}
\usage{
remove_duplicates(de_novos)
}
\arguments{
\item{de_novos}{dataframe of variants}
}
\value{
dataframe of variants, without duplicates
}
\description{
The autism de novos come from multiple studies. Some probands are in multiple
studies, so we have to remove duplicate entries, while accounting for
potential differences in sample IDs, and possibly different coordinates (for
indels).
}
\examples{
variants = read.table(header=TRUE, text="
    person_id    chrom start_pos ref_allele alt_allele consequence        study_phenotype
    sample_01    1     10000     AAA        A          frameshift_variant autism
    sample_01.p1 1     10001     AAA        A          frameshift_variant normal_iq_autism
    sample_02    1     10001     AAA        A          frameshift_variant normal_iq_autism",
    stringsAsFactors=FALSE)
remove_duplicates(variants)
}

