% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_sanders_de_novos.R
\name{sanders_de_novos}
\alias{sanders_de_novos}
\title{get de novo data from the Sanders et al autism exome study}
\usage{
sanders_de_novos()
}
\value{
data frame of de novos, with standardised genome coordinates and VEP
}
\description{
Supplementary table 2 (where the excel sheets for the probands and
siblings have been combined) from:
Sanders et al. (2012) Nature 485:237-241
doi: 10.1038/nature10945
}

