% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_zaiidi_de_novos.R
\name{zaidi_de_novos}
\alias{zaidi_de_novos}
\title{get de novo data from Zaidi et al. congenital heart disease exome study}
\usage{
zaidi_de_novos()
}
\value{
data frame of de novos, including gene symbol, functional consequence
    (VEP format), chromosome, nucleotide position and SNV or INDEL type
}
\description{
De novo mutation data sourced from Supplementary table 4:
Zaidi et al. (2013) Nature 498:220-223
doi: 10.1038/nature12141
}

